Update 1
I've added the WSA logos to the webpage and as a favicon as well as a title and short introduction.

I've updated the category objects to include a "total" class, which records the total amount of funding each category has received overall. The subtitle to each pie chart has been changed to display the total.

There's a new function, dollarString, which takes that int total and turns it into a string with a $ and comma in appropriate places; this makes it a little easier to read.

Example usage:
dollarString(106000) ==> "$106,000"